Moved windows test into IntegrationTests
