--- Assets/Tests/InputSystem/WindowsTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d07ecd3cab9a99e44b1d0150d4b608ab
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
